\name{untb-package}
\alias{untb-package}
\alias{untb}
\docType{package}
\title{
\packageTitle{untb}
}
\description{
\packageDescription{untb}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{untb}
\packageIndices{untb}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{untb}

Maintainer: \packageMaintainer{untb}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
